Factor out `vec_reserve` to reduce system call invocations
`vec_reserve`, a newly introduced macro is used in `vec_init_with`.
This reduces the number of `realloc` calls, since we reserve enough
space in advance. Now we can simply do one `memcpy` to copy all the
arguments we need. This is possible because we can calculate the
number of __VA_ARGS__ ahead-of-time.